requirements.txt
beautifulsoup4==4.11.1
black==22.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
idna==3.4
mypy-extensions==0.4.3
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
pathspec==0.10.3
platformdirs==2.6.0
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
